CHORE: rename action_modules to actions
