+++ device_manager/actions/camera_actions.py
